requirements.txt
asgiref==3.3.4
cx-Freeze==6.2
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.11.0
django-heroku==0.3.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2==2.9.3
SQLAlchemy==1.4.26
sqlparse==0.2.4
whitenoise==5.3.0
